API/requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
astroid==3.3.8
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
databases==0.9.0
Deprecated==1.2.18
dill==0.3.9
fastapi==0.115.8
flake8==7.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
isort==6.0.0
limits==4.0.1
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
pylint==3.3.4
pytest==8.3.4
python-dotenv==1.0.1
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tomlkit==0.13.2
typing_extensions==4.12.2
uvicorn==0.34.0
wrapt==1.17.2
